Rename component QuoteCarousel
