+++ src/components/QuoteCarousel.js
